% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfGMM-package.R
\docType{package}
\name{cfGMM-package}
\alias{cfGMM}
\alias{cfGMM-package}
\title{cfGMM: closed-form estimation for gamma mixture model}
\description{
This package is for function cfGMM, which gives estimate of fitted gamma mixture model.
}
\keyword{internal}
