% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbk.R
\name{optimize_derivative}
\alias{optimize_derivative}
\title{Function with the constraint}
\usage{
optimize_derivative(bk, mk, x, phi.vec)
}
\arguments{
\item{bk}{scale}

\item{mk}{median}

\item{x}{data vector}

\item{phi.vec}{phi}
}
\description{
Function with the constraint
}
