% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfGMM.R
\name{cfGMM}
\alias{cfGMM}
\title{Main Function:cfGMM}
\usage{
cfGMM(
  x,
  k,
  alpha = NULL,
  beta = NULL,
  lambda = NULL,
  diff.conv = 1e-06,
  max.iter = 1000,
  maxrestarts = 20,
  max.comp = FALSE,
  min.lambda = 1e-04,
  constraint = NULL,
  inner.loop = 4
)
}
\arguments{
\item{x}{data vector}

\item{k}{number of components. Interger that is greater or equal to 2}

\item{alpha}{(optional) shape starting value. Can be either a vector with length k or a constant. If the input is vector and there is NA in the vector, the NAs will be filled with MOM estimator. If the input is a constant, the initial value of a will be a vector of length k. If not specified, the initial value will be a vector of values near MOM estimator. All numerical values input should be greater than 0.}

\item{beta}{(optional) scale starting value. Same specification as init.val.a}

\item{lambda}{(optional) Can be either a matrix with nrow=length(xtor) and ncol=k, or a vector with length k. NA value matrix input is not allowed. If the input is vector, the initial value would be a matrix with nrow=length(xtor) and repeated rows of the vector input. If the input is not specified, the initial phi matrix is a nrow=length(xtor) and ncol=k with all elements equal to 1/k.}

\item{diff.conv}{(optional) default to 1e-5}

\item{max.iter}{(optional) default to 1e3}

\item{maxrestarts}{(optional) default to 20}

\item{max.comp}{(optional) whether k indicates maximum number of components or fixed number of component. Logical, TRUE indicates k is the maximum and FALSE indicates k is fixed. Default to be FALSE.}

\item{min.lambda}{1e-5}

\item{constraint}{NULL or matrix of dimension (k,2), each row correspond to alpha_lower, alpha_upper, beta_lower, beta_upper for one component. If one end is not specified, let it be NA.}

\item{inner.loop}{integer greater or equal to 1. When there isn't constraint, the more inner.loop, the more accurate the result, the longer the algorithm takes. Defaults to be 4.}
}
\description{
Main Function:cfGMM
}
